cast value to gint for g_print.

+++ libbalsa/mailbox_mbox.c
@@ -559,7 +559,7 @@ lbm_mbox_restore(LibBalsaMailboxMbox * mbox)
 #ifdef DEBUG
     g_print("%s: %s file has %d messages\n", __func__,
             LIBBALSA_MAILBOX(mbox)->name,
-            length / sizeof(struct message_info));
+            (gint) (length / sizeof(struct message_info)));
 #endif
 
     msg_info = (struct message_info *) contents;
@@ -598,7 +598,7 @@ lbm_mbox_restore(LibBalsaMailboxMbox * mbox)
 #ifdef DEBUG
     g_print("%s: %s restored %d messages\n", __func__,
             LIBBALSA_MAILBOX(mbox)->name,
-            msg_info - (struct message_info *) contents);
+            (gint) (msg_info - (struct message_info *) contents));
 #endif
 
     mbox_stream = mbox->gmime_stream;
